user info on angular created. update on sql database

+++ OutboundREST/src/main/java/com/skilldistillery/outbound/services/HuntTripServicesImpl.java
@@ -7,7 +7,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.skilldistillery.outbound.entities.hunttrip.HuntTrip;
+import com.skilldistillery.outbound.entities.user.User;
 import com.skilldistillery.outbound.repository.HuntTripRepository;
+import com.skilldistillery.outbound.repository.UserRepository;
 
 
 @Service
@@ -16,38 +18,49 @@ public class HuntTripServicesImpl implements HuntTripServices {
 	@Autowired
 	private HuntTripRepository huntRepo;
 
+	@Autowired
+	private UserRepository userRepo;
 	
 	@Override
-	public List<HuntTrip> findAllHuntTrips() {
-		return huntRepo.findAll();
+	public List<HuntTrip> findAllHuntTrips(String username) {
+		return huntRepo.findByUser_Username(username);
 	}
 
 
 	@Override
-	public HuntTrip findById(int huntId) {
+	public HuntTrip findById(String username,int huntId) {
 		Optional<HuntTrip> huntOp = huntRepo.findById(huntId);
+		User user = userRepo.findByUsername(username);
 		HuntTrip hunt = null;
+		if(user !=null) {
 		if(huntOp.isPresent()) {
 			hunt = huntOp.get();
+			hunt.setUser(user);
 		}
 		
+		}
 		return hunt;
 	}
 
 
 	@Override 
-	public HuntTrip createHuntingTrip(HuntTrip hunt) {
-		
+	public HuntTrip createHuntingTrip(String username,HuntTrip hunt) {
+		User user = userRepo.findByUsername(username);
+		if(user!= null) {
+			hunt.setUser(user);
+		}
 		return huntRepo.saveAndFlush(hunt);
 	}
 
  
 
 	@Override
-	public boolean deleteHuntTrip(int huntId) {
+	public boolean deleteHuntTrip(String username, int huntId) {
 		boolean deleted = false;
+		
 		Optional<HuntTrip> huntOp = huntRepo.findById(huntId);
-		if(huntOp.isPresent()) {
+		
+		if(huntOp.isPresent() && huntOp.get().getUser().getUsername().equals(username)) {
 			huntRepo.deleteById(huntId);
 			deleted=true;
 			
@@ -58,13 +71,12 @@ public class HuntTripServicesImpl implements HuntTripServices {
 
 
 	@Override
-	public HuntTrip updateHuntingTrip(HuntTrip hunt, int huntId) {
+	public HuntTrip updateHuntingTrip(String username, HuntTrip hunt, int huntId) {
 		hunt.setId(huntId);
-		if(huntRepo.existsById(huntId)) {
-			return huntRepo.save(hunt);
-		}
+		User user = userRepo.findByUsername(username);
+		hunt.setUser(user);
 	
-		return null; 
+		return huntRepo.save(hunt); 
 	}
